Add more async chain test, move asynh chain back into one file
JS / TS didnt like the Abstract Class loop we had with the chaining
setup."TypeError: Class extends value undefined is not a constructor or null "
